requirements.txt
# Core Flask dependencies
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.45
flask-cors==6.0.2
PyJWT==2.10.1
Werkzeug==3.1.3

# PDF Generation
reportlab==4.2.5
weasyprint==62.3

# Testing
pytest==9.0.2
hypothesis==6.148.8

# Development
python-dotenv==1.0.0